revert ipfs renaming
