+++ core/blockchain/impl/weight_calculator_impl.hpp
@@ -8,7 +8,7 @@
 
 #include "blockchain/weight_calculator.hpp"
 
-#include "storage/ipfs/ipfs_datastore.hpp"
+#include "storage/ipfs/datastore.hpp"
 
 namespace fc::blockchain::weight {
   enum class WeightCalculatorError { kNoNetworkPower = 1 };
