+++ core/storage/ipfs/impl/ipfs_block_service.hpp
@@ -8,7 +8,7 @@
 
 #include <memory>
 
-#include "storage/ipfs/ipfs_datastore.hpp"
+#include "storage/ipfs/datastore.hpp"
 
 namespace fc::storage::ipfs {
   class IpfsBlockService : public IpfsDatastore {
