+++ core/storage/ipfs/merkledag/impl/merkledag_service_impl.hpp
@@ -8,7 +8,7 @@
 
 #include <memory>
 
-#include "storage/ipfs/ipfs_datastore.hpp"
+#include "storage/ipfs/datastore.hpp"
 #include "storage/ipfs/merkledag/impl/leaf_impl.hpp"
 #include "storage/ipfs/merkledag/merkledag_service.hpp"
 #include "storage/ipld/ipld_link.hpp"
