+++ test/testutil/mocks/storage/ipfs/ipfs_datastore_mock.hpp
@@ -8,7 +8,7 @@
 
 #include <gmock/gmock.h>
 
-#include "storage/ipfs/ipfs_datastore.hpp"
+#include "storage/ipfs/datastore.hpp"
 
 namespace fc::storage::ipfs {
